Code Review :)
This looks like you've a solid understanding of the session and hidden variable concepts in this assignment. There is room for improvement. Session should be used as minimally as possible, only holding information that needs to persist through the whole user-session. Things like color or the amount of gold generated each request should be done elsewhere. Things like an activity message log and the gold earned would be things to keep in session. I would suggest looking into a different way to display the activity messages, if you need help figuring out ways to do this, DM me on slack. Hint: session['activity_log'] = [ ]